+++ packages/terra-compact-interactive-list/src/utils/keyHandlerUtils.js
@@ -53,28 +53,17 @@ export const moveFocusFromElement = (containerRef, id, moveForward) => {
  */
 export const handleDownKey = (focusedCell, numberOfColumns, flowHorizontally, rowsLength) => {
   const { row, cell } = focusedCell;
-  if (flowHorizontally) {
-    const visualColumnNumber = row % numberOfColumns;
-    const nextRow = row + numberOfColumns;
-    if (nextRow > rowsLength - 1) {
-      if (visualColumnNumber === numberOfColumns - 1) {
-        // the last semantic row in the last visual column, focus shouldn't go anywhere
-        return focusedCell;
-      }
-      // the last semantic row in current visual column, focus should move to the next visual column
-      return { row: visualColumnNumber + 1, cell };
+  const visualColumnNumber = row % numberOfColumns;
+  const nextRow = row + numberOfColumns;
+  if (nextRow > rowsLength - 1) {
+    if (visualColumnNumber === numberOfColumns - 1) {
+      // the last semantic row in the last visual column, focus shouldn't go anywhere
+      return focusedCell;
     }
-    return { row: nextRow, cell };
+    // the last semantic row in current visual column, focus should move to the next visual column
+    return { row: visualColumnNumber + 1, cell };
   }
-  // vertical flow
-  const rowsPerColumn = Math.ceil(rowsLength / numberOfColumns);
-  const numberOfPlaceholders = (rowsPerColumn * numberOfColumns) - rowsLength;
-  if ((numberOfPlaceholders === 0 && row + 1 === rowsLength) || (numberOfPlaceholders !== 0 && row + 1 === rowsLength + numberOfPlaceholders - 1)) {
-    // the last semantic row in the last visual column, focus shouldn't go anywhere
-    return focusedCell;
-  }
-  // focus should go to the next semantic row unless the last row
-  return { row: row + 1 < rowsLength ? row + 1 : row, cell };
+  return { row: nextRow, cell };
 };
 
 /**
@@ -92,21 +81,17 @@ export const handleUpKey = (focusedCell, numberOfColumns, flowHorizontally, rows
     return focusedCell;
   }
   const rowsPerColumn = Math.ceil(rowsLength / numberOfColumns);
-  if (flowHorizontally) {
-    const visualColumnNumber = row % numberOfColumns;
-    if (row === visualColumnNumber) {
-      // the first semantic row in current visual column, focus should move to the previous visual column last semantic row
-      let newRow = (numberOfColumns * rowsPerColumn) - (numberOfColumns - (visualColumnNumber - 1));
-      if (newRow > rowsLength - 1) {
-        newRow -= numberOfColumns;
-      }
-      return { row: newRow, cell };
+  const visualColumnNumber = row % numberOfColumns;
+  if (row === visualColumnNumber) {
+    // the first semantic row in current visual column, focus should move to the previous visual column last semantic row
+    let newRow = (numberOfColumns * rowsPerColumn) - (numberOfColumns - (visualColumnNumber - 1));
+    if (newRow > rowsLength - 1) {
+      newRow -= numberOfColumns;
     }
-    // moving focus ove visual row up
-    return { row: row - numberOfColumns, cell };
+    return { row: newRow, cell };
   }
-  // vertical flow
-  return { row: row - 1, cell };
+  // moving focus one visual row up
+  return { row: row - numberOfColumns, cell };
 };
 
 /**
@@ -121,13 +106,7 @@ const getFirstSemanticRowIndexInVisualRow = (rowsLength, numberOfColumns, flowHo
   if (row === undefined || row === null) {
     return 0;
   }
-  if (flowHorizontally) {
-    const firstItemInVisualRow = (Math.floor(row / numberOfColumns) * numberOfColumns);
-    return firstItemInVisualRow;
-  }
-  const rowsPerColumn = Math.ceil(rowsLength / numberOfColumns);
-  const visualColumnNumber = Math.floor(row / rowsPerColumn);
-  const firstItemInVisualRow = row - (visualColumnNumber * rowsPerColumn);
+  const firstItemInVisualRow = (Math.floor(row / numberOfColumns) * numberOfColumns);
   return firstItemInVisualRow;
 };
 
@@ -167,7 +146,7 @@ export const handleLeftKey = (event, focusedCell, numberOfColumns, flowHorizonta
     }
     // The first cell. Focus moves to the last cell of the semantic row to the left.
     nextCell = cellsLength - 1;
-    nextRow -= flowHorizontally ? 1 : Math.ceil(rowsLength / numberOfColumns);
+    nextRow -= 1;
     return { row: nextRow, cell: nextCell };
   }
   // Not first cell. Focus moves within the row to the next cell to the left.
@@ -187,14 +166,8 @@ const getLastSemanticRowIndexInVisualRow = (rowsLength, numberOfColumns, flowHor
     // If current row omitted, return the index of the last element
     return rowsLength - 1;
   }
-  if (flowHorizontally) {
-    const lastItemInVisualRow = (Math.ceil((row + 1) / numberOfColumns) * numberOfColumns) - 1;
-    return lastItemInVisualRow < rowsLength - 1 ? lastItemInVisualRow : rowsLength - 1;
-  }
-  const rowsPerColumn = Math.ceil(rowsLength / numberOfColumns);
-  const rowsToTop = row % rowsPerColumn;
-  const lastItemInVisualRow = (numberOfColumns - 1) * rowsPerColumn + rowsToTop;
-  return lastItemInVisualRow < rowsLength ? lastItemInVisualRow : lastItemInVisualRow - rowsPerColumn;
+  const lastItemInVisualRow = (Math.ceil((row + 1) / numberOfColumns) * numberOfColumns) - 1;
+  return lastItemInVisualRow < rowsLength - 1 ? lastItemInVisualRow : rowsLength - 1;
 };
 
 /**
@@ -241,7 +214,7 @@ export const handleRightKey = (event, focusedCell, numberOfColumns, flowHorizont
     }
     // Focus moves to the first cell of the next semantic row.
     nextCell = 0;
-    nextRow += flowHorizontally ? 1 : Math.ceil(rowsLength / numberOfColumns);
+    nextRow += 1;
   } else {
     // Focus moves to the next cell to the right in the same semantic row.
     nextCell += 1;