+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples.2/CellContent.jsx
@@ -140,7 +140,6 @@ const CellContent = () => {
       onCellSelect={onCellSelect} // a callback function for cell selection
       onClearSelection={onClearSelection}
       rowHeaderIndex={1}
-      rowHeight="50px"
     />
   );
 };